requirements.txt
torch==2.3.0
torchvision==0.18.0
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
scikit-learn==1.5.0
onnx==1.16.0
onnxruntime==1.17.1